File length limit (Stream.CopyToLimited).

+++ src/Services.Core/IResponseLimiter.cs
@@ -7,18 +7,18 @@ namespace SenseNet.Services.Core
         /// <summary>
         /// Gets the maximum file length in bytes.
         /// </summary>
-        int MaxFileLengthInBytes { get; }
+        long MaxFileLengthInBytes { get; }
         /// <summary>
         /// Gets the maximum response length in bytes.
         /// </summary>
-        int MaxResponseLengthInBytes { get; }
+        long MaxResponseLengthInBytes { get; }
 
         /// <summary>
         /// If the passed value exceeds the configured MaxFileLengthInBytes limit,
         /// ApplicationException will be thrown.
         /// </summary>
         /// <param name="fileLength">The length of file to write.</param>
-        void AssertFileLength(int fileLength);
+        void AssertFileLength(long fileLength);
 
         /// <summary>
         /// Aggregates the "textLength" using the httpContext as a background storage.
@@ -27,7 +27,7 @@ namespace SenseNet.Services.Core
         /// </summary>
         /// <param name="httpContext">The current HttpContext instance.</param>
         /// <param name="textLength">The length of the text to write.</param>
-        void AssertResponseLimit(HttpContext httpContext, int textLength);
+        void AssertResponseLimit(HttpContext httpContext, long textLength);
 
         /// <summary>
         /// Returns all written bytes. Designed only test purposes.
